#3 Green light refactoring
