+++ SQLGeneration/Builders/DDL/IDatabaseObject.cs
@@ -6,7 +6,7 @@ using System.Text;
 namespace SQLGeneration.Builders
 {
     /// <summary>
-    /// Specifies any object that can be defined in the database that can be subjected to DDL statements such as "CREATE", "ALTER" and "DROP"
+    /// Specifies any object that can be defined in the database and subjected to DDL statements such as "CREATE", "ALTER" and "DROP" statements.
     /// </summary>
     public interface IDatabaseObject : IVisitableBuilder
     {
