+++ SQLGeneration/SQLGeneration.csproj
@@ -52,12 +52,16 @@
     <Compile Include="Builders\BuilderVisitor.cs" />
     <Compile Include="Builders\ConditionalCaseBranch.cs" />
     <Compile Include="Builders\ConditionalCase.cs" />
+    <Compile Include="Builders\DDL\AutoIncrement.cs" />
     <Compile Include="Builders\DDL\ColumnDefinition.cs" />
-    <Compile Include="Builders\DDL\ColumnSize.cs" />
+    <Compile Include="Builders\DDL\ColumnDefinitionList.cs" />
+    <Compile Include="Builders\DDL\DataType.cs" />
     <Compile Include="Builders\DDL\Database.cs" />
+    <Compile Include="Builders\DDL\DefaultConstraint.cs" />
     <Compile Include="Builders\DDL\ICreateBuilder.cs" />
     <Compile Include="Builders\DDL\IDatabaseObject.cs" />
     <Compile Include="Builders\DDL\CreateBuilder.cs" />
+    <Compile Include="Builders\DDL\TableDefinition.cs" />
     <Compile Include="Builders\PrecedingUnboundFrame.cs" />
     <Compile Include="Builders\FollowingUnboundFrame.cs" />
     <Compile Include="Builders\FollowingBoundFrame.cs" />
